Don't use Comic Sans for the inflation card
I don't seem to have Comic Sans available so it would end up using a
different font which made the text too long and it would crop off the
end of the card. This changes it to use a free font called Komika
Display, but also converts the text object to a path so that you don't
need to have the font installed to render it.

http://www.fontsquirrel.com/fonts/Komika-Display